references/manim-to-video-claw/manimtovideo/requirements.txt
fastapi==0.108.0
httpx==0.28.1
pydantic==2.10.6
python-dotenv>=0.21.0,<0.22.0
regex==2024.11.6
uvicorn==0.25.0
oss2==2.18.4
manim==0.19.0
manim-voiceover[azure]
baidu-aip==4.16.13
chardet==5.2.0
pandas
scipy
sympy
matplotlib
aiofiles
aiohttp==3.11.12
uvloop
runpod
supabase==2.15.0